COMMERCE-6188 return int like persistence instead of long

+++ modules/apps/account/account-service/src/main/java/com/liferay/account/service/impl/AccountEntryUserRelLocalServiceImpl.java
@@ -323,7 +323,7 @@ public class AccountEntryUserRelLocalServiceImpl
 	}
 
 	@Override
-	public long getAccountEntryUserRelsCountByAccountEntryId(
+	public int getAccountEntryUserRelsCountByAccountEntryId(
 		long accountEntryId) {
 
 		return accountEntryUserRelPersistence.countByAEI(accountEntryId);
